+++ eval.js
@@ -1,1 +1,0 @@
-export { hydraEval } from './src/eval.js'